requirements.txt
streamlit>=1.31.0
langchain>=0.1.12
langchain-community>=0.0.28
langchain-core>=0.1.32
chromadb>=0.4.24
sentence-transformers>=2.5.1
python-dotenv>=1.0.1
requests>=2.31.0
